Unlink Service By Proposal Refactoring

+++ src/modules/proposals/services/UnlinkServiceByProposalService.ts
@@ -1,5 +1,7 @@
 import { inject, injectable } from "tsyringe";
 
+import { AppError } from "@shared/errors/AppError";
+
 import { IServicesProposalsRepository } from "../repositories/IServicesProposalsRepository";
 
 @injectable()
@@ -10,6 +12,12 @@ export class UnlinkServiceByProposalService {
     private serviceProposalRepository: IServicesProposalsRepository
   ) { }
   public async execute(service_proposal_id: string): Promise<void> {
+    const serviceProposalAlreadyExists = await this.serviceProposalRepository.listServicesProposalById(service_proposal_id);
+
+    if (!serviceProposalAlreadyExists) {
+      throw new AppError("Service proposal not found");
+    }
+
     const serviceProposal = await this.serviceProposalRepository.delete(service_proposal_id);
 
     return serviceProposal;
